mvp-platform/apps/data-service/requirements.txt
Flask==2.3.2
flask-cors==4.0.0
kafka-python==2.0.2
redis==4.6.0
psycopg2-binary==2.9.7
prometheus-flask-exporter==0.22.4
requests==2.31.0